+++ packages/core/src/modules/comment.ts
@@ -4,7 +4,7 @@ import { mergeBorder } from "./cell";
 import { Context, getFlowdata } from "../context";
 import { CellMatrix, GlobalCache } from "../types";
 import { colLocation, rowLocation } from "./location";
-import { isAllowEdit } from "../utils";
+// import { isAllowEdit } from "../utils";
 
 export function getArrowCanvasSize(
   fromX: number,
@@ -273,8 +273,8 @@ export function newComment(
   // if(!checkProtectionAuthorityNormal(Store.currentSheetId, "editObjects")){
   //     return;
   // }
-  const allowEdit = isAllowEdit(ctx);
-  if (!allowEdit) return;
+  // const allowEdit = isAllowEdit(ctx);
+  // if (!allowEdit) return;
   if (ctx.hooks.beforeInsertComment?.(r, c) === false) {
     return;
   }
@@ -318,8 +318,8 @@ export function editComment(
   // if(!checkProtectionAuthorityNormal(Store.currentSheetId, "editObjects")){
   //     return;
   // }
-  const allowEdit = isAllowEdit(ctx);
-  if (!allowEdit) return;
+  // const allowEdit = isAllowEdit(ctx);
+  // if (!allowEdit) return;
   const flowdata = getFlowdata(ctx);
   removeEditingComment(ctx, globalCache);
   const comment = flowdata?.[r][c]?.ps;
@@ -347,8 +347,8 @@ export function deleteComment(
   // if(!checkProtectionAuthorityNormal(Store.currentSheetId, "editObjects")){
   //     return;
   // }
-  const allowEdit = isAllowEdit(ctx);
-  if (!allowEdit) return;
+  // const allowEdit = isAllowEdit(ctx);
+  // if (!allowEdit) return;
   if (ctx.hooks.beforeDeleteComment?.(r, c) === false) {
     return;
   }
@@ -622,7 +622,7 @@ export function onCommentBoxResize(
   globalCache: GlobalCache,
   e: MouseEvent
 ) {
-  if (ctx.allowEdit === false) return false;
+  // if (ctx.allowEdit === false) return false;
   const commentBox = globalCache?.commentBox;
   if (commentBox?.resizingId && commentBox.resizingSide) {
     const box = document.getElementById(commentBox.resizingId);
@@ -716,8 +716,8 @@ export function onCommentBoxMove(
   globalCache: GlobalCache,
   e: MouseEvent
 ) {
-  const allowEdit = isAllowEdit(ctx);
-  if (!allowEdit) return false;
+  // const allowEdit = isAllowEdit(ctx);
+  // if (!allowEdit) return false;
   const commentBox = globalCache?.commentBox;
   if (commentBox?.movingId) {
     const box = document.getElementById(commentBox.movingId);